Move exception in Ant UI tests to correct package
The TestAgainException is placed in a downstream package of the Ant UI
tests but still used in upstream packages. This change properly moves
the exception to a central package where is it is primarily referenced.
