+++ ant/org.eclipse.ant.tests.ui/test plugin/org/eclipse/ant/tests/ui/testplugin/TestAgainExceptionRule.java
@@ -11,7 +11,6 @@
  *******************************************************************************/
 package org.eclipse.ant.tests.ui.testplugin;
 
-import org.eclipse.ant.tests.ui.debug.TestAgainException;
 import org.junit.rules.TestRule;
 import org.junit.runner.Description;
 import org.junit.runners.model.Statement;
